Fixed authentication Bugs Added Sign in & Guest User Added Events and news to home page

+++ دعوات قطر/SignInViewController.h
@@ -10,7 +10,7 @@
 
 @interface SignInViewController : UIViewController <UITextFieldDelegate>
 
-@property (nonatomic) int userID;
+@property (nonatomic) NSInteger userID;
 
 @property (weak, nonatomic) IBOutlet UITextField *mobileField;
